Added starting/stopping selenium to integration test process.
NOTE: 
The integration tests do not appear to fail when there actually should be a failure.

There are bugs in the current version of selenium that prevent execution on OS X 10.5+. 

Execution on Windows 7 also appears to work the first time, but fail subsequent times (Selenium isn't closing the browser). Once again this is a known bug.


git-svn-id: http://google-sitebricks.googlecode.com/svn/trunk@61 c7392d14-1f41-11de-a108-cd2f117ce590