Embellishments to stat publishing: 1) adding mapbinder-based configuration of StatsServlet  > this facilitates application-specific additions of publishers
 2) publishing a richer snapshot of data
  > moves from Map<String, String> to Map<StatDescriptor, Object>
  > this gives each publisher the name, description, and java value of a stat
  
 3) adding tests for the new publishers
  
  Note: the description field is yet unused, but it's clear how it
  can now be used by a publisher, versus having been hidden before.

Dependency changes:
* this CL migrates from google-collections to guava
* this CL introduces a dependency on google gson

IML files:
Note, because I'm using IJ 10, I'm not sure that my local
edits are going to be backwards compatible.  I left them out of this
CL.  This means that the guava and gson deps will need to be added
to the .iml files for intellij to pick up the new deps (I'm not sure what
the best thing to do here would have been).

git-svn-id: http://google-sitebricks.googlecode.com/svn/trunk@196 c7392d14-1f41-11de-a108-cd2f117ce590